Update FavoriteMedia.plugin.js
I tried and i got it to a somewhat working version, this version is not fully tested, but i figured I'd publish it since a lot of people have been asking for this plugin, and quite frankly it is very useful and I'd hate to see it go abandoned. 
However, I'm by no means attempting to take over the development for this plugin.

Few points to keep in mind, there was one single module that I'm yet to fix so if anyone else wants to have a go at it, and also if anyone wants to update the filters and make them more resilient.